Initial changes

+++ src/main/java/com/toolsrental/demo/dto/ToolsRentalResponseDTO.java
@@ -25,7 +25,7 @@ public class ToolsRentalResponseDTO {
 
     private double dailyRentalCharge;
 
-    private int chargeDays;
+    private long chargeDays;
 
     private double preDiscountCharge;
